requirements_macos_silicon.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
comtypes==1.2.0
docopt==0.6.2
docx==0.2.4
enum-compat==0.0.3
et-xmlfile==1.1.0
filetype==1.2.0
frozenlist==1.4.0
greenlet==3.0.3
hachoir==3.0a3
HTMLParser==0.0.2
hupper==1.12
idna==3.4
Jinja2==3.1.2
line-profiler==4.1.1
line-profiler-pycharm==1.1.0
lxml==5.1.0
MarkupSafe==2.1.3
multidict==6.0.4
olefile==0.46
openpyxl==3.1.2
pandas==2.2.0
PasteDeploy==3.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
player==0.6.1
psutil==5.9.5
pycaw==20230407
pycparser==2.21
pyramid==2.0.2
pyramid-jinja2==2.10
python-dateutil==2.8.2
python-docx==1.1.0
python-pptx==0.6.23
python-vlc==3.0.18122
requests==2.31.0
six==1.16.0
texttable==1.6.7
toml==0.10.2
translationstring==1.4
typing_extensions==4.9.0
tzdata==2023.4
unicodecsv==0.14.1
urllib3==2.0.5
venusian==3.0.0
vulture==2.9.1
WebOb==1.8.7
xlrd==2.0.1
XlsxWriter==3.1.9
yarg==0.1.9
yarl==1.9.2
numpy==1.26.3
chardet==5.2.0
python-magic==0.4.27
pdf2image==1.17.0
moviepy==1.0.3
PyPDF2==3.0.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
Pillow==10.4.0
pytsk3==20231007
libewf-python==20240506
python-registry==1.3.1
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
opencv-python==4.10.0.84
